src/requirements.txt
apturl==0.5.2
asgiref==3.5.0
astroid==2.5.6
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0
django-countries==7.2.1
django-phone-field==1.8.1
duplicity==0.8.12.0
entrypoints==0.3
fabric==2.5.0
fasteners==0.14.1
feedparser==6.0.10
filelock==3.4.2
flake8==4.0.1
flake8-polyfill==1.0.2
Flask==2.1.1
Flask-Babel==2.0.0
Flask-HTTPAuth==4.5.0
future==0.18.2
gpg===1.13.1-unknown
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.3
iniconfig==1.1.1
invoke==1.3.0
isort==5.9.1
itsdangerous==2.1.2
Jinja2==3.1.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
mccabe==0.6.1
monotonic==1.5
mysqlclient==2.1.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
packaging==21.3
paramiko==2.6.0
pep8==1.7.1
pep8-naming==0.12.1
pexpect==4.6.0
Pillow==8.2.0
platformdirs==2.4.1
pluggy==1.0.0
protobuf==3.6.1
py==1.11.0
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
pyflakes==2.4.0
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.8.3
pymacaroons==0.13.0
pymongo==4.1.1
PyNaCl==1.3.0
pyparsing==3.0.8
PyQt5==5.14.1
pyRFC3339==1.1
PySocks==1.6.8
pytest==7.1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
redis==4.3.4
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
sgmllib3k==1.0.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.4.32
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
toml==0.10.2
tomli==2.0.1
torbrowser-launcher==0.3.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.13.0
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==2.1.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.8.0
